requirements.txt
pandas
SQLAlchemy
tabulate
numpy
Flask
streamlit
python-dotenv
scikit-learn
